Bot-discord-eeggcc/requirements.txt
discord.py>=2.0
numpy
openpyxl
requests
Pillow